refactor(front: wants/types): ディレクトリ名の変更
